+++ components/Post/PostImage.tsx
@@ -1,0 +1,3 @@
+export function PostImage() {
+  return null;
+}
